// apps/indexer/src/database/schemas/public/Balance.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { type ColumnType, type Selectable, type Insertable, type Updateable } from 'kysely';

/** Identifier type for public.Balance */
export type BalanceId = string & { __brand: 'BalanceId' };

/** Represents the table public.Balance */
export default interface BalanceTable {
  id: ColumnType<BalanceId, BalanceId | undefined, BalanceId>;

  userAddress: ColumnType<string, string, string>;

  tokenAddress: ColumnType<string, string, string>;

  balance: ColumnType<bigint, bigint, bigint>;
}

export type Balance = Selectable<BalanceTable>;

export type NewBalance = Insertable<BalanceTable>;

export type BalanceUpdate = Updateable<BalanceTable>;